Created TestCase for common loading file features, treating encoding file formats
